Crate epic_keychain

Library module for the key holder functionalities provided by Epic.

Re-exports

pub use crate::extkey_bip32::ChildNumber;

pub use crate::keychain::ExtKeychain;

pub use crate::view_key::ViewKey;

Modules

extkey_bip32

Implementation of BIP32 hierarchical deterministic wallets, as defined at https://github.com/bitcoin/bips/blob/master/bip-0032.mediawiki Modified from above to integrate into epic and allow for different hashing algorithms if desired

keychain

mnemonic

BIP39 Implementation

view_key

Structs

BlindSum

Accumulator to compute the sum of blinding factors. Keeps track of each factor as well as the “sign” with which they should be combined.

BlindingFactor

ExtKeychainPath

Encapsulates a max 4-level deep BIP32 path, which is the most we can currently fit into a rangeproof message. The depth encodes how far the derivation depths go and allows differentiating paths. As m/0, m/0/0 or m/0/0/0/0 result in different derivations, a path needs to encode its maximum depth.

Identifier

Enums

Error

SwitchCommitmentType

Constants

IDENTIFIER_SIZE

Traits

Keychain