Module epic_keychain::mnemonic

BIP39 Implementation

Implementation of BIP39 Mnemonic code for generating deterministic keys, as defined at https://github.com/bitcoin/bips/blob/master/bip-0039.mediawiki

Structs

WORDS

List of bip39 words

Enums

Error

An error that might occur during mnemonic decoding

Functions

from_entropy

Converts entropy to a mnemonic

search

Returns the index of a word in the wordlist

to_entropy

Converts a mnemonic to entropy

to_seed

Converts a nemonic and a passphrase into a seed